+++ fl_image_picker/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2
+
+* Fixed an issue with `SingleImagePicker` not showing
+* Error callback changed to enum
+
 ## 0.1.1
 
 * Fix errors on the web